+++ django_s3_storage/conf.py
@@ -54,6 +54,10 @@ class LazySettings(object):
         name = "AWS_S3_BUCKET_NAME",
     )
 
+    AWS_S3_KEY_PREFIX = LazySetting(
+        name = "AWS_S3_KEY_PREFIX",
+    )
+
     AWS_S3_BUCKET_AUTH = LazySetting(
         name = "AWS_S3_BUCKET_AUTH",
         default = True,
@@ -70,13 +74,17 @@ class LazySettings(object):
         name = "AWS_S3_BUCKET_NAME_STATIC",
     )
 
+    AWS_S3_KEY_PREFIX_STATIC = LazySetting(
+        name = "AWS_S3_KEY_PREFIX_STATIC",
+    )
+
     AWS_S3_BUCKET_AUTH_STATIC = LazySetting(
-        name = "AWS_S3_BUCKET_AUTH",
+        name = "AWS_S3_BUCKET_AUTH_STATIC",
         default = False,
     )
 
     AWS_S3_MAX_AGE_SECONDS_STATIC = LazySetting(
-        name = "AWS_S3_MAX_AGE_SECONDS",
+        name = "AWS_S3_MAX_AGE_SECONDS_STATIC",
         default = 60 * 60 * 24 * 365,  # 1 year.
     )